refactor: 네이밍 변경(HouseStructure -> FloorPlan)
