+++ src/pages/onboarding/ImageGenerationFunnel.tsx
@@ -1,11 +1,11 @@
 import { useImageGenerationFunnel } from './hooks/useImageGenerationFunnel.hooks';
 import Step1HouseInfo from './components/steps/step1/Step1HouseInfo';
-import Step2HouseStructure from './components/steps/step2/Step2HouseStructure';
+import Step2FloorPlan from './components/steps/step2/Step2FloorPlan';
 import Step3InteriorTaste from './components/steps/step3/Step3InteriorTaste';
 import Step4MainActivity from './components/steps/step4/Step4MainActivity';
 import {
   type CompletedHouseInfo,
-  type CompletedHouseStructure,
+  type CompletedFloorPlan,
   type CompletedInteriorTaste,
 } from './types/funnel';
 import FunnelLayout from './components/layout/FunnelLayout';
@@ -19,7 +19,7 @@ export const ImageGenerationFunnel = () => {
         HouseInfo={funnel.Render.with({
           events: {
             selectHouseInfo: (data: CompletedHouseInfo, { history }) => {
-              history.push('HouseStructure', data);
+              history.push('FloorPlan', data);
             },
           },
           render({ dispatch, context }) {
@@ -31,20 +31,17 @@ export const ImageGenerationFunnel = () => {
             );
           },
         })}
-        HouseStructure={funnel.Render.with({
+        FloorPlan={funnel.Render.with({
           events: {
-            selectHouseStructure: (
-              data: CompletedHouseStructure,
-              { history }
-            ) => {
+            selectedFloorPlan: (data: CompletedFloorPlan, { history }) => {
               history.push('InteriorTaste', data);
             },
           },
           render({ dispatch, context }) {
             return (
-              <Step2HouseStructure
+              <Step2FloorPlan
                 context={context}
-                onNext={(data) => dispatch('selectHouseStructure', data)}
+                onNext={(data) => dispatch('selectedFloorPlan', data)}
               />
             );
           },
