+++ src/pages/onboarding/components/steps/step2/Step2FloorPlan.css.ts
